wip: rename type
