+++ src/routing/deprecated-decorator/route-metadata.ts
@@ -1,6 +1,6 @@
 import type { HttpMiddleware } from '@/Http';
 import type { RouteOptions } from '@/routing/route-options';
-import type { RouterMethod } from '@/routing/router-method';
+import type { RouterMethod } from '@/routing/router-method.type';
 
 /**
  * @deprecated Legacy decorator routing metadata. Use `@koala-ts/framework/routing` instead.
